requirements.txt
asyncpraw==7.2.0
beautifulsoup4~=4.9.3
python-dotenv~=0.17.1
discord.py==1.7.2
dropbox==11.11.0